+++ index.js
@@ -7,8 +7,7 @@ const {promisify} = require('util')
 const postcss = require('postcss')
 const got = require('got')
 const checkSvg = require('is-svg')
-const fileType = require('file-type')
-const pMap = require('p-map')
+const FileType = require('file-type')
 const debug = require('debug')
 
 const _log = debug('b64:log')
@@ -17,12 +16,13 @@ const _info = debug('b64:opt')
 
 const readFile = promisify(_readFile)
 const urlRegx = /^https?:\/\//
-const b64Regx = /b64-{3}["']?([\w.\-/:]+)["']?-{3}/gm
+const b64Regx = /(?<match_>b64-{3}["']?(?<file_>[\w.\-/:]+)["']?-{3})/gm
+const declsRegx = /^background(-image)?$|^src$/
 
 async function _find(dir, file) {
-	_log('_find ---> ', file)
+	_log('_find | file ---> ', file)
 	if (urlRegx.test(file)) {
-		const {body} = await got(file, {encoding: null, retries: 2, timeout: 5000})
+		const body = await got(file, {retries: 2, timeout: 5000}).buffer()
 		return body
 	}
 
@@ -30,44 +30,51 @@ async function _find(dir, file) {
 	return readFile(f)
 }
 
-function _mime(buf) {
+async function _mime(buf) {
 	const isSvg = checkSvg(buf.toString('utf-8'))
 	if (isSvg) {
 		return 'image/svg+xml'
 	}
 
-	const {mime} = fileType(buf)
+	const {mime} = await FileType.fromBuffer(buf)
 	return mime
 }
 
 async function _inline(dir, file) {
 	const buf = await _find(dir, file)
-	const mime = _mime(buf)
+	const mime = await _mime(buf)
 	return `data:${mime};charset=utf-8;base64,${buf.toString('base64')}`
 }
 
-module.exports = postcss.plugin('postcss-inline-base64', (opts = {}) => (css, result) => {
-	const {to = process.cwd()} = result.opts
-	const options = {...{baseDir: dirname(to)}, ...opts}
-	_info(options)
-	const inlines = []
-	css.walkDecls(/^background(-image)?$|^src$/, decl => {
-		const matches = decl.value.match(b64Regx) || []
-		for (const match of matches) {
-			const file = match.replace(b64Regx, '$1')
-			inlines.push({file, match, decl})
-		}
-	})
-	return pMap(inlines, async ({file, match, decl}) => {
-		const node = decl.parent
-		let data = file
-		try {
-			data = await _inline(options.baseDir, file)
-		} catch (error) {
-			node.warn(result, error.message)
-			_error(error.message)
-		} finally {
-			decl.value = decl.value.replace(match, data)
-		}
-	})
-})
+async function _parse(...args) {
+	const [file, match, decl, options, result] = args
+	const node = decl.parent
+	let data = file
+	try {
+		data = await _inline(options.baseDir, file)
+	} catch (error) {
+		node.warn(result, error.message)
+		_error('_parse | error ---> ', error.message)
+	} finally {
+		decl.value = decl.value.replace(match, data)
+	}
+}
+
+function _callback(options_ = {}) {
+	return (css, result) => {
+		const {to = process.cwd()} = result.opts
+		const options = {...{baseDir: dirname(to)}, ...options_}
+		_info('_callback | options ---> ', options)
+		const inlines = new Set([])
+		css.walkDecls(declsRegx, decl => {
+			let matches
+			while ((matches = b64Regx.exec(decl.value)) !== null) {
+				const {file_, match_} = matches.groups
+				inlines.add(_parse(file_, match_, decl, options, result))
+			}
+		})
+		return Promise.allSettled(inlines)
+	}
+}
+
+module.exports = postcss.plugin('postcss-inline-base64', _callback)